# src/swagger/swagger-config/swagger.yml
openapi: 3.0.0
paths:
...
  /menu/filter:
    get:
      operationId: Menucontroller_getMenuByName
      parameters:
        - name: itemName
          required: false
          in: query
          schema:
            type: string
        - name: category
          required: false
          in: query
          schema:
            type: string
      responses:
        '200':
          description: ''
      tags: *ref_3
  /menu:
    post:
      operationId: Menucontroller_createMenu
      summary: ''
      description: Creates a new menuitem category
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MenuDto'
      responses:
        '201':
          description: ''
      tags: *ref_3
  /menu/{id}/menuitem:
    post:
      operationId: Menucontroller_addMenuItem
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MenuItemDto'
      responses:
        '200':
          description: Creates a new Menuitem
          content:
            application/json:
              examples:
                success:
                  description: Creates a new Menu
                  value:
                    succes: true
                    data:
                      menu_itemname: Pancakes
                      price: 20
                      menus:
                        menu_id: 1
                        menu_name: lunch
                        date:
                          createdDate: '2023-08-17T08:58:58.467Z'
                          updatedDate: '2023-08-17T08:58:58.467Z'
                      menuitem_id: 17
                      tax: 5
                      date:
                        createdDate: '2023-08-25T11:20:06.064Z'
                        updatedDate: '2023-08-25T11:20:06.064Z'
        '400':
          description: Invalid id
          content:
            application/json:
              examples:
                MenuItemIdFound:
                  description: Occurs when Menuid is not found
                  value:
                    message: Given_id_is_not_found
      tags: *ref_3
  /employee/login:
    post:
      operationId: AuthController_employeeLogin
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/loginEmployeeDto'
      responses:
        '200':
          description: Logins the user and returns a jwt token after verifying the user
          content:
            application/json:
              examples:
                succes:
                  description: Verify the employee and returns a Jwt token
                  value:
                    succes: true
                    jwtToken: sdjhvshvjdvnjdvnjvkjsvvhjvvjjjeiljejnebjq
        '400':
          description: Errors in given request body
          content:
            application/json:
              examples:
                MenuItemIdFound:
                  description: Occurs when given Id is invalid
                  value:
                    message: Given email is already registered give new email
      tags: &ref_4
        - EmployeeAuth
  /employee/register:
    post:
      operationId: AuthController_employeeRegister
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/registerEmployeeDto'
      responses:
        '200':
          description: Logins the user and returns a jwt token after verifying the user
          content:
            application/json:
              examples:
                succes:
                  description: Verify the employee and returns a Jwt token
                  value:
                    succes: true
                    jwtToken: sdjhvshvjdvnjdvnjvkjsvvhjvvjjjeiljejnebjq
        '400':
          description: Errors in given request body
          content:
            application/json:
              examples:
                MenuItemIdFound:
                  description: Occurs when given Id is invalid
                  value:
                    message: Given email is already registered give new email
        '401':
          description: Unauthorized
          content:
            application/json:
              examples:
                AcceswithoutAuthorization:
                  description: ' If the client attempts to access certain book-related endpoints without providing any authentication credentials'
                  value:
                    message: Authentication_is_required_to_access_the_resource.
                ExpiredAuthenticationToken:
                  description: >-
                    When the client sends an authentication token that is
                    invalid, expired, or not recognized by the server
                  value:
                    message: client's_token_is_not_valid_for_the_requested_operation
        '403':
          description: Forbidden
          content:
            application/json:
              examples:
                InsufficientPrivileges:
                  description: ' If the client provides authentication credentials, but the credentials do not grant sufficient privileges to access a specific book-related operation'
                  value:
                    message: >-
                      client_does_not_have_the_required_authorization_to_perform_the_requested_action.
                RoleBasedAccessControl:
                  description: >-
                    If the API uses role-based access control, the server may
                    respond with a 403 status code when a user with a certain
                    role tries to access an operation that requires a higher
                    role or specific permissions.
                  value:
                    message: >-
                      User_with_this_role_is_not_permitted_to_access_the_resource
      tags: *ref_4
  /employee/{id}/details:
    get:
      operationId: AuthController_getEmployeeDetails
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: number
      responses:
        '200':
          description: Gives the order details by given id
          content:
            application/json:
              examples:
                success:
                  description: Gives the menu details
                  value:
                    succes: true
                    data:
                      menuitem_id: 2
                      menu_itemname: manchuria
                      price: 10
                      tax: 5
                      date:
                        createdDate: '2023-08-17T08:31:36.802Z'
                        updatedDate: '2023-08-24T04:40:27.390Z'
        '400':
          description: Occurs when given Id is invalid
          content:
            application/json:
              examples:
                InvalidId:
                  description: Occurs when given Id is invalid
                  value:
                    message: Given_id_is_not_found
      tags: *ref_4
  /employee/roles:
    post:
      operationId: AuthController_createRoles
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/createRoleDto'
      responses:
        '200':
          description: creates a role
          content:
            application/json:
              examples:
                success:
                  description: Gives the role deatils that have been posted
                  value:
                    succes: true
                    data:
                      name: Customer
                      description: customer has permission to order
                      id: 5
        '400':
          description: Occurs when you are trying to post the same role again
          content:
            application/json:
              examples:
                conflict:
                  description: Same Role posted again
                  value:
                    error: Conflict Error
                    message: 'Given_role_is_already_present '
        '401':
          description: Unauthorized
          content:
            application/json:
              examples:
                AcceswithoutAuthorization:
                  description: ' If the client attempts to access certain book-related endpoints without providing any authentication credentials'
                  value:
                    message: Authentication_is_required_to_access_the_resource.
                ExpiredAuthenticationToken:
                  description: >-
                    When the client sends an authentication token that is
                    invalid, expired, or not recognized by the server
                  value:
                    message: client's_token_is_not_valid_for_the_requested_operation
        '403':
          description: Forbidden
          content:
            application/json:
              examples:
                InsufficientPrivileges:
                  description: ' If the client provides authentication credentials, but the credentials do not grant sufficient privileges to access a specific book-related operation'
                  value:
                    message: >-
                      client_does_not_have_the_required_authorization_to_perform_the_requested_action.
                RoleBasedAccessControl:
                  description: >-
                    If the API uses role-based access control, the server may
                    respond with a 403 status code when a user with a certain
                    role tries to access an operation that requires a higher
                    role or specific permissions.
                  value:
                    message: >-
                      User_with_this_role_is_not_permitted_to_access_the_resource
      tags: *ref_4
  /employee/role/{id}:
    put:
      operationId: AuthController_updateRoles
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/updateRoleDto'
      responses:
        '200':
          description: Role is updated succefully
          content:
            application/json:
              examples:
                success:
                  description: Role is updated with given name in database
                  value:
                    succes: true
                    data:
                      id: 1
                      employee_Id: A1
                      employee_Name: prem
                      status: Active
                      email: prem3a3
                      phoneNumber: 8367
                      date:
                        createdDate: '2023-08-22T05:30:40.783Z'
                        updatedDate: '2023-08-22T09:38:24.000Z'
                      roles:
                        id: 1
                        name: Manager
        '400':
          description: Occurs when given path varaible is invalid
          content:
            application/json:
              examples:
                IdNotFound:
                  description: Role is updated with given name in database
                  value:
                    message: Id not found to update the role
      tags: *ref_4
info:
  title: chandu
  description: Your API Description
  version: '1.0'
  contact: {}
...
servers:
  - url: http://localhost:3000
components:
  securitySchemes:
    bearer:
      scheme: bearer
      bearerFormat: JWT
      type: http
  schemas:
...
    MenuDto:
      type: object
      properties:
        menu_name:
          type: string
          example: Starters
      required:
        - menu_name
    MenuItemDto:
      type: object
      properties:
        menu_itemname:
          type: string
          example: Pancakes
        price:
          type: number
          example: 20
      required:
        - menu_itemname
        - price
    IsEmail:
      type: object
      properties: {}
    loginEmployeeDto:
      type: object
      properties:
        email:
          example: prem@gmail.com
          allOf:
            - $ref: '#/components/schemas/IsEmail'
        password:
          type: isStrongPassword
          example: prem@123
      required:
        - email
        - password
    registerEmployeeDto:
      type: object
      properties:
        employee_Id:
          type: string
          example: A23
        employee_Name:
          type: string
          example: A23
        status:
          type: string
          example: Active
        email:
          type: string
          example: prem3a3@gmail.com
        password:
          type: string
          example: Prem@12345678
        role:
          type: string
          example: Manager
      required:
        - employee_Id
        - employee_Name
        - status
        - email
        - password
        - role
    createRoleDto:
      type: object
      properties:
        description:
          type: string
          example: Have permissions to change an employee
      required:
        - description
    updateRoleDto:
      type: object
      properties:
        name:
          type: string
          example: Manager
        description:
          type: string
          example: Have permissions to change an employee
      required:
        - name
        - description